+++ src/test_app.c
@@ -15,7 +15,7 @@ int main(void) {
 
     if(cir_error_count(ir) > 0) {
         fprintf(stderr, "\n--- ERRORS ---\n");
-        for (uint64_t i = 0; i < cir_error_count(ir); i++) {
+        for (size_t i = 0; i < cir_error_count(ir); i++) {
             fprintf(stderr, "%s\n", cir_get_error(ir, i));
         }
         fprintf(stderr, "\n");
